% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_excel.R
\name{edit_excel}
\alias{edit_excel}
\title{Edit table with Excel}
\usage{
edit_excel(x)
}
\arguments{
\item{x}{table; object to edit}
}
\description{
Open/edit a file using desktop version of excel  q q
}
\details{
Works on Windows only.

System has to be set up so that Excel reads CSV files. This is generally, the
default.
}
\examples{

\dontrun{
  df <- edit_excel(iris)
}

}
\references{
\url{http://stackoverflow.com/a/12164899/1504321}
}
\seealso{
edit, edit.data.frame
}
